chatbot_backend/requirements.txt
aiofiles==24.1.0
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.7.8
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.109.0
filelock==3.19.1
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
openpyxl==3.1.2
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.8.0
pypdf==3.17.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
redis==5.0.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
PyMySQL==1.1.1
cryptography==43.0.3
sentence-transformers==2.2.2
starlette==0.35.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
xlsxwriter==3.2.9
